# library-system/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "hhahaaaahahahahaha",
        "title": "图书管理系统 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/borrow": {
            "post": {
                "description": "用户借阅指定图书（需要登录）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "borrow"
                ],
                "summary": "借阅图书",
                "parameters": [
                    {
                        "description": "借书信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.BorrowBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "借书成功",
                        "schema": {
                            "$ref": "#/definitions/handlers.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户未认证",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "图书不存在",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "库存不足或借阅次数已达上限",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/borrow/records": {
            "get": {
                "description": "获取当前用户的所有借阅记录（需要登录）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "borrow"
                ],
                "summary": "获取用户借阅记录",
                "responses": {
                    "200": {
                        "description": "借阅记录数组",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.BorrowRecord"
                            }
                        }
                    },
                    "401": {
                        "description": "用户未认证",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/borrow/return": {
            "post": {
                "description": "用户归还已借阅的图书（需要登录）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "borrow"
                ],
                "summary": "归还图书",
                "parameters": [
                    {
                        "description": "还书信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ReturnBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "还书成功",
                        "schema": {
                            "$ref": "#/definitions/handlers.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户未认证",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "借阅记录或图书不存在",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "图书已归还",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.BorrowBookRequest": {
            "type": "object",
            "required": [
                "book_id"
            ],
            "properties": {
                "book_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ReturnBookRequest": {
            "type": "object",
            "required": [
                "record_id"
            ],
            "properties": {
                "record_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "handlers.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "操作成功"
                }
            }
        },
...
        "models.BorrowRecord": {
            "type": "object",
            "properties": {
                "book_id": {
                    "type": "integer",
                    "example": 1
                },
                "borrowed_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "due_date": {
                    "type": "string",
                    "example": "2024-02-15T10:30:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "returned_at": {
                    "type": "string",
                    "example": "2024-01-18T09:15:00Z"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "用户登录后，Session Cookie会自动携带在请求中",
            "type": "apiKey",
            "name": "library-session",
            "in": "cookie"
        }
    }
}